Split common ACE_Settings.hpp from config.cpp, shortened sampleSetting
